+++ pareidolia-desktop/package.json
@@ -4,10 +4,11 @@
   "version": "1.0.0",
   "description": "My Electron application description",
   "main": ".vite/build/main.js",
+  "type": "module",
   "private": true,
   "scripts": {
     "start": "electron-forge start",
-    "cloud": "electron --headless .",
+    "cloud-server": "node src/server.js",
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
